move useJoinedEvents into useEvents
